--- src/main/java/acc/firewatch/member/entity/Member.java
@@ -1,33 +1,0 @@
-package acc.firewatch.member.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @Column(nullable = false, unique = true, length = 20)
-    private String phoneNum;
-
-    @Setter
-    @Column(nullable = false)
-    private String password;
-
-    @Embedded
-    private Address address;
-
-    @Column(nullable = false)
-    private boolean verified;
-
-}
-
